@@ -4,28 +4,21 @@
   "type": "library",
   "license": "GPL-2.0-or-later",
   "authors": [
-	{
-	  "name": "Daniel Mejta",
-	  "email": "daniel@mejta.net"
-	}
+    {
+      "name": "Daniel Mejta",
+      "email": "daniel@mejta.net"
+    }
   ],
   "autoload": {
-	"psr-4": {
-	  "Wpify\\Model\\": "src/"
-	}
+    "psr-4": {
+      "Wpify\\Model\\": "src/"
+    }
   },
   "scripts": {
-	"generate-unit-test": "codecept generate:test unit",
-	"test": "codecept run unit --debug"
   },
   "require": {
-	"php": ">=7.3.0||^8.0.0",
-	"phpstan/phpdoc-parser": "^0.5.6"
+    "php": ">=8.0.0"
   },
   "require-dev": {
-	"codeception/codeception": "^4.1",
-	"codeception/module-phpbrowser": "^1.0.0",
-	"codeception/module-asserts": "^1.0.0",
-    "brain/monkey": "^2.6"
   }
 }
