@@ -1,31 +1,26 @@
 <?php
 
+declare( strict_types=1 );
+
 namespace Wpify\Model;
 
-use Wpify\Model\Abstracts\AbstractOrderItemModel;
+use Wpify\Model\Attributes\AccessorObject;
 
-/**
- * Class Order
- * @package Wpify\Model
- * @property ProductRepository $_repository
- * @method \WC_Order_Item_Shipping source_object()
- */
-class OrderItemShipping extends AbstractOrderItemModel {
-	public $method_id;
-	public $instance_id;
-	public $method_title;
+class OrderItemShipping extends OrderItem {
 	/**
-	 * @var string[][]
+	 * Method ID.
 	 */
-	protected $_props = array(
-		'id'           => array( 'source' => 'object', 'source_name' => 'id' ),
-		'type'         => array( 'source' => 'object', 'source_name' => 'type' ),
-		'name'         => array( 'source' => 'object', 'source_name' => 'name' ),
-		'quantity'     => array( 'source' => 'object', 'source_name' => 'quantity' ),
-		'tax_total'    => array( 'source' => 'object', 'source_name' => 'tax_total' ),
-		'tax_class'    => array( 'source' => 'object', 'source_name' => 'tax_class' ),
-		'method_id'    => array( 'source' => 'object', 'source_name' => 'method_id' ),
-		'instance_id'  => array( 'source' => 'object', 'source_name' => 'instance_id' ),
-		'method_title' => array( 'source' => 'object', 'source_name' => 'method_title' ),
-	);
+	#[AccessorObject]
+	public string $method_id = '';
+
+	/**
+	 * Instance ID.
+	 */
+	#[AccessorObject]
+	public string $instance_id = '';
+	/**
+	 * Method title.
+	 */
+	#[AccessorObject]
+	public string $method_title = '';
 }
