@@ -1,27 +1,36 @@
 <?php
+declare( strict_types=1 );
 
 namespace Wpify\Model;
 
-use Wpify\Model\Abstracts\AbstractPostRepository;
+use Wpify\Model\Interfaces\ModelInterface;
 
 /**
- * Class BasePageRepository
- *
- * @package Wpify\Model
- *
- * @method Page[] all()
- * @method Page[] find( array $args = array() )
- * @method Page create()
- * @method Page get( $object = null )
- * @method mixed save( $model )
- * @method mixed delete( $model )
+ * @method Page|null get( mixed $source )
+ * @method Page create( array $data )
+ * @method Page save( Page $model )
+ * @method bool delete( Page $model )
+ * @method Page[] find( array $args = [] )
+ * @method Page[] find_all( array $args = array() )
+ * @method Page[] find_published( array $args = array() )
+ * @method Page[] find_all_by_term( Page $model )
  */
-class PageRepository extends AbstractPostRepository {
-	static function post_type(): string {
-		return 'page';
-	}
-
+class PageRepository extends PostRepository {
+	/**
+	 * Returns the model class name.
+	 *
+	 * @return string
+	 */
 	public function model(): string {
 		return Page::class;
+	}
+
+	/**
+	 * Returns post types for the repository.
+	 *
+	 * @return string[]
+	 */
+	public function post_types(): array {
+		return array( 'page' );
 	}
 }
