@@ -1,27 +1,16 @@
 <?php
+declare( strict_types=1 );
 
 namespace Wpify\Model;
 
-use Wpify\Model\Abstracts\AbstractTermModel;
-use Wpify\Model\Interfaces\PostModelInterface;
-use Wpify\Model\Relations\TermPostsRelation;
+use Wpify\Model\Attributes\TermPostsRelation;
 
-/**
- * Class Category
- * @package Wpify\Model
- *
- * @method CategoryRepository model_repository()
- */
-class Category extends AbstractTermModel {
-
-	/** @var PostModelInterface */
-	public $posts;
-
-	protected function posts_relation(): TermPostsRelation {
-		return new TermPostsRelation(
-			$this,
-			'posts',
-			$this->model_repository()->get_post_repository()
-		);
-	}
+class Category extends Term {
+	/**
+	 * Posts assigned to this category.
+	 *
+	 * @var Post[]
+	 */
+	#[TermPostsRelation( Post::class )]
+	public array $posts = array();
 }
