@@ -1,102 +1,104 @@
 <?php
+declare( strict_types=1 );
 
 namespace Wpify\Model;
 
-use Wpify\Model\Abstracts\AbstractTermRepository;
+use Wpify\Model\Exceptions\RepositoryNotFoundException;
+use Wpify\Model\Exceptions\RepositoryNotInitialized;
 
 /**
- * Class BasePageRepository
+ * Repository for the menu.
  *
- * @package Wpify\Model
- *
+ * @method Menu create( array $data )
+ * @method Menu save( Menu $model )
+ * @method bool delete( Menu $model )
+ * @method Menu[] find( array $args )
+ * @method Menu[] find_all()
+ * @method Menu[] find_not_empty( array $args = array() )
+ * @method Menu[] find_children_of( int $parent_id = 0, array $args = array() )
+ * @method Menu[] find_terms_of_post( int $post_id )
  */
-class MenuRepository extends AbstractTermRepository {
+class MenuRepository extends TermRepository {
+	public const RETURN_ARRAY = 'array';
+	public const RETURN_OBJECTS = 'objects';
+
+	private array $menu_locations = array();
+
+	/**
+	 * Returns the model class name.
+	 *
+	 * @return string
+	 */
 	public function model(): string {
 		return Menu::class;
 	}
 
-	protected function factory( $object ) {
-		$class = $this->model();
-
-		return new $class( $this->resolve_object( $object ), $this );
-	}
-
-	protected function resolve_object( $data = null ): ?\WP_Term {
-		$menu_id   = null;
-		$locations = get_nav_menu_locations();
-		if ( $data != 0 && is_numeric( $data ) ) {
-			$menu_id = $data;
-		} elseif ( is_array( $locations ) && ! empty( $locations ) ) {
-			$menu_id = $this->get_menu_id_from_locations( $data, $locations );
-		} elseif ( $data === false ) {
-			$menu_id = false;
-		}
-		if ( ! $menu_id ) {
-
-			$menu_id = $this->get_menu_id_from_terms( $data );
-		}
-
-		return parent::resolve_object( $menu_id );
-	}
-
 	/**
-	 * @param string $slug
-	 * @param array  $locations
+	 * Returns taxonomy for the repository.
 	 *
-	 * @return integer
-	 * @internal
+	 * @return string
 	 */
-	protected function get_menu_id_from_locations( $slug, $locations ) {
-		if ( $slug === 0 ) {
-			$slug = $this->get_menu_id_from_terms( $slug );
-		}
-		if ( is_numeric( $slug ) ) {
-			$slug = array_search( $slug, $locations );
-		}
-		if ( isset( $locations[ $slug ] ) ) {
-			$menu_id = $locations[ $slug ];
-			if ( function_exists( 'wpml_object_id_filter' ) ) {
-				$menu_id = wpml_object_id_filter( $locations[ $slug ], 'nav_menu' );
-			}
-
-			return $menu_id;
-		}
-	}
-
-	/**
-	 * @param int $slug
-	 *
-	 * @return int
-	 * @internal
-	 */
-	protected function get_menu_id_from_terms( $slug = 0 ) {
-		if ( ! is_numeric( $slug ) && is_string( $slug ) ) {
-			// we have a string so lets search for that
-			$menu = get_term_by( 'slug', $slug, 'nav_menu' );
-			if ( $menu ) {
-				return $menu->term_id;
-			}
-			$menu = get_term_by( 'name', $slug, 'nav_menu' );
-			if ( $menu ) {
-				return $menu->term_id;
-			}
-		}
-		$menus = get_terms( 'nav_menu', array( 'hide_empty' => true ) );
-		if ( is_array( $menus ) && count( $menus ) ) {
-			if ( isset( $menus[0]->term_id ) ) {
-				return $menus[0]->term_id;
-			}
-		}
-
-		return 0;
-	}
-
-	public function get_menu_item_repository() {
-		// TODO: cache this
-		return new MenuItemRepository();
-	}
-
 	public function taxonomy(): string {
 		return 'nav_menu';
+	}
+
+	/**
+	 * Returns the Menu model by the location, WP_Term object, id, slug or name.
+	 *
+	 * @param mixed $source
+	 *
+	 * @return Menu|null
+	 * @throws Exceptions\RepositoryNotInitialized
+	 * @throws RepositoryNotFoundException
+	 */
+	public function get( mixed $source ): ?Menu {
+		if ( empty( $this->menu_locations ) ) {
+			$this->menu_locations = get_nav_menu_locations();
+		}
+
+		$wp_menu     = null;
+		$model_class = $this->model();
+
+		if ( is_string( $source ) && in_array( $source, array_keys( $this->menu_locations ) ) ) {
+			$wp_menu = wp_get_nav_menu_object( $this->menu_locations[ $source ] );
+		}
+
+		if ( $wp_menu ) {
+			$menu = parent::get( $wp_menu );
+		} else {
+			$term = parent::get( $source );
+			$menu = $term ? new $model_class( $this->manager() ) : null;
+
+			$menu->source( $term->source() );
+		}
+
+		return $menu;
+	}
+
+	/**
+	 * Returns the menu items by location, WP_Term object, id, slug or name.
+	 *
+	 * @param mixed $source Menu location, WP_Term object, id, slug or name.
+	 * @param string $return Return type, can be MenuRepository::RETURN_OBJECTS or MenuRepository::RETURN_ARRAY.
+	 *
+	 * @return array|MenuItem[]
+	 * @throws RepositoryNotInitialized
+	 * @throws RepositoryNotFoundException
+	 */
+	public function items( mixed $source, string $return = self::RETURN_OBJECTS ): array {
+		$items = array();
+		$menu  = $this->get( $source );
+
+		if ( $menu ) {
+			if ( $return === self::RETURN_OBJECTS ) {
+				return $menu->children;
+			} elseif ( $return === self::RETURN_ARRAY ) {
+				$menu = $menu->to_array();
+
+				return $menu['children'];
+			}
+		}
+
+		return $items;
 	}
 }
