@@ -1,12 +1,9 @@
 <?php
 
+declare( strict_types=1 );
+
 namespace Wpify\Model;
 
-/**
- * Class Order
- * @package Wpify\Model
- * @property ProductRepository $_repository
- * @method \WC_Order_Item_Fee source_object()
- */
-class OrderItemFee extends \Wpify\Model\Abstracts\AbstractOrderItemModel {
+class OrderItemFee extends OrderItem {
+
 }
